Added support for own DNS configuration.
Before it commit, the tool tried to use system configuration and if can't use it then fails when used with the -i option because can not resolve the IPs. Now if the tool fails to get system DNS configuration it can use their own implementation of DNS resolvers in the following order: Quad9, Cloudflare, Google.