Ask to crt.sh only for subdomains
When we pass to crt.sh API `%example.com`, we are also asking for certificates that include `example.com`, after it commit we only ask for subdomains passing `%.example.com` as parameter (note the dot between % and example.com). As result, the API query's are most faster and we doesn't have to waste time to delete unneeded results inside the application when calling `sort()` and `dedup()` methods.