Use a static name for output files.
Before it commit Findomain was using a random name for writing to output files, now the output file name is just the name of the target. It mean that you can run Findomain with the -o flag and automatize your workflow easily.

A example of that is just: `findomain -t example.com -o && cat example.com | httprobe` in order to use the httprobe tool.

Fix #12.