Use HashSets instead of Vecs.
It's a planed change, HashSets are faster than Vecs at the cost of more memory but in the case of Findomain that's irrelevant due the very low resources usage of the tool.

Hashsets make easy to work with the new subdomains as you can see in the commented code. It's the first step for the implementation of #13.