Revert self-updater but keep updates checker.
It's a change in order to keep Findomain transparency with the users. Self updater basically was forcing the user to download a precompiled binary (even if the user has compiled form source), that isn't the right way to do the things. The updates checker is still there and can be accessed using the --check-updates option. It will tell the user if there's or not available updates.